Remove auto-updated modified column on Mindset
The SQLAlchemy docs describe a method for automatically generating UTC
timestamps at [1] but it fails with an error “construct has no default
compilation handler.”. The docs don’t describe a way of defining a
default compilation handler. As this functionality is not that
important right now, I am disabling it.

[1]:
http://docs.sqlalchemy.org/en/rel_1_0/core/compiler.html#utc-timestamp-f
unction